processing_script/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
face-recognition>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.31.0
